Moved transaction-related methods to `TransactionMetics`
